refactor(frontend): move prelude composition into starlark
