+++ crates/once-frontend/prelude/index.star
@@ -1,0 +1,5 @@
+PRELUDE_SOURCES = [
+    "common.star",
+    "apple.star",
+    "rust.star",
+]
